+++ internal/acceptance_tests/blocks/acl_explicit_on_compute.tf
@@ -1,0 +1,5 @@
+resource "fastly_service_cdn_acl" "test" {
+  service_id = fastly_service_compute.test.id
+  version    = {{.SERVICE_VERSION}}
+  name       = "{{.ACL_NAME}}"
+}