+++ internal/acceptance_tests/blocks/dictionary_explicit_write_only.tf
@@ -1,0 +1,7 @@
+resource "fastly_service_dictionary" "test" {
+  service_id    = fastly_service_cdn.test.id
+  version       = {{.SERVICE_VERSION}}
+  name          = "{{.DICTIONARY_NAME}}"
+  write_only    = true
+  force_destroy = true
+}